extract extension module
